implement maxcost using ids and limit as costs
