+++ server/game/baseability.js
@@ -2,7 +2,6 @@ const _ = require('underscore');
 const AbilityTargetCard = require('./AbilityTargets/AbilityTargetCard.js');
 const AbilityTargetRing = require('./AbilityTargets/AbilityTargetRing.js');
 const AbilityTargetSelect = require('./AbilityTargets/AbilityTargetSelect.js');
-const Costs = require('./costs.js');
 /**
  * Base class representing an ability that can be done by the player. This
  * includes card actions, reactions, interrupts, playing a card, marshaling a
@@ -25,9 +24,6 @@ class BaseAbility {
     constructor(properties) {
         this.cost = this.buildCost(properties.cost);
         this.targets = this.buildTargets(properties);
-        if(properties.max) {
-            this.cost.push(Costs.playMax());
-        }
     }
 
     buildCost(cost) {
@@ -188,7 +184,7 @@ class BaseAbility {
     }
 
     isCardAbility() {
-        return true;
+        return false;
     }
 }
