+++ server/game/basecard.js
@@ -338,7 +338,7 @@ class BaseCard {
     }
     
     canTriggerAbilities(context) {
-        return this.allowGameAction('triggerAbilities', context);
+        return !this.facedown && this.allowGameAction('triggerAbilities', context);
     }
     
     getMenu() {
